requirements.txt
pyrebase
stockstats
pandas-datareader
keras
tensorflow
sklearn
numpy==1.19.5
jwt
